+++ source/Nuke.Common/Tools/DotCover/DotCoverTasks.cs
@@ -3,31 +3,17 @@
 // https://github.com/nuke-build/nuke/blob/master/LICENSE
 
 using System;
-using System.Linq;
-using JetBrains.Annotations;
-using Nuke.Core;
 using Nuke.Core.Tooling;
 
 namespace Nuke.Common.Tools.DotCover
 {
-    public static partial class DotCoverTasks
+    partial class DotCoverTasks
     {
-        private static string GetPackageExecutable()
+        [Obsolete("Use " + nameof(DotCoverAnalyseSettings) + "." + nameof(DotCoverAnalyseSettingsExtensions.SetTargetSettings) + " instead.")]
+        public static void DotCoverAnalyse(Action testAction, Configure<DotCoverAnalyseSettings> configurator = null, ProcessSettings processSettings = null)
         {
-            throw new NotImplementedException();
-        }
-        
-        [CanBeNull]
-        private static IProcess StartProcess(DotCoverAnalyseSettings toolSettings, ProcessSettings processSettings)
-        {
-            var testAction = toolSettings.TestAction.NotNull("testAction != null");
-            var capturedStartInfo = ProcessTasks.CaptureProcessStartInfo(testAction);
-            toolSettings = toolSettings
-                .SetTargetExecutable(capturedStartInfo.ToolPath)
-                .SetTargetArguments(capturedStartInfo.Arguments)
-                .SetTargetWorkingDirectory(capturedStartInfo.WorkingDirectory);
-
-            return ProcessTasks.StartProcess(toolSettings, processSettings);
+            configurator = configurator ?? (x => x);
+            DotCoverAnalyse(x => configurator(x).SetTestAction(testAction));
         }
     }
 }